+++ src/Typings/enums.ts
@@ -46,6 +46,6 @@ export enum QuestionTypes {
 export enum Routes {
   Category = "https://opentdb.com/api_count.php?",
   Questions = "https://opentdb.com/api.php?",
-  SessionReset = "",
+  SessionReset = "https://opentdb.com/api_token.php?command=reset",
   SessionStart = "https://opentdb.com/api_token.php?command=request",
 }